e2etest/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.49.1",
    "pytest-playwright>=0.6.2",
    "python-dotenv>=1.0.1",
]
